+++ pingpongcallc/include/ping.h
@@ -5,6 +5,6 @@
 typedef int64_t custom_int;
 #endif
 
-void user_do_ping_spg(int *v);
-void user_ping_spg(int i);
+void user_do_ping_spg(long long *v);
+void user_ping_spg(long long i);
 void recover(void);
